More cleanup or unecessary code and components. Removed starter package since this is not a server app, and does not have plugins, so we don't require any dynamic loading. Removed ClassUtils.class due to the same reason. Removed the src/bin dir and bash script stubs since we likely won't be installing anything on a linux server. Removed the client listener since we dont' have plugins that would require to be notified when the client was shutting down. Merged the Client Info class and interface since it was redundant.
