--- src/bin/install
@@ -1,42 +1,0 @@
-#!/usr/bin/env bash
-#
-# Narvaro: @VERSION@
-# Build Date: @DATE@
-# Commit Head: @HEAD@
-# JDK: @JDK@
-# ANT: @ANT@
-#
-# Installation Script
-#
-# This script installs Narvaro on the server.
-#
-# Run this script as root after unpacking Narvaro tar.gz or zip file.
-# It is expected that you are executing this script from the bin directory.
-#
-
-# Fail on any error
-set -e
-
-# Check that we are root ... so non-root users stop here.
-if [ "$(id -u)" != "0" ]; then
-    echo "${0} must be run as root"
-    exit 1
-fi
-
-OLD_PWD="$(pwd)"
-
-# Ensure we are running from the bin/ directory of the
-# unpacked narvaro archive
-if [[ ! "${OLD_PWD}" =~ narvaro/bin ]]; then
-    echo "${0} must be run from the extracted narvaro bin/ directory"
-    exit 1
-fi
-
-#########################################
-# Do setup stuff here for database, etc #
-#########################################
-
-# Source the postinstall script and execute it
-. postinstall
-
-echo "Narvaro installation complete"
